Detect and remove self intersections of QQuadPath
Currently the CurveRenderer is not working when the path is
self-intersecting.

With this patch, the self-intersections are removed before the
path is used for filling (optionally, default: on) The stroking
path is untouched.

The function findOverlappingCandidates finds candidates of elements
that might be intersecting. Its complexity is O(n log n) and can
also be used in other parts of the code where overlapping bounding
triangles need to be identified.

The function solveIntersections removes all intersections from
a QQuadPath. If intersections are solved, the path is oriented such
that the filling is on the right side of the path. If no intersections
are found, the path is returned without any changes. The optional
argument alwaysReorder can be used to force a reordering of the paths,
such that the filling of the shape is always on the right side of the
path.

Intersections are found with Newtons algorithm with 9 different
starting values. This is reliable in finding all intersections but
the starting values could be improved/reduced to improve performance.

Change-Id: I088e4edfff755155521ed91114bc67f63c6e546a
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
(cherry picked from commit 06e42e733ed6658abbb30ba7be2e571b4533d009)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>