Add setter and notifier for QQuickTextDocument's document
Users want to be able to provide their own "special" instances of
QTextDocument, or custom subclasses. But so far we allow this only
from C++, not from QML.

QQuickTextDocument is created lazily in QQuickTextEdit::textDocument()
so it's not possible to rely on it to hold the actual QTextDocument all
the time. It's only a facade provided to the user. But QQuickTextEdit
holds the QTextDocument, and is the only class that creates
QQuickTextDocument, and sets the parent to itself; so QQuickTextDocument
can rely on accessing the TextEdit's document rather than maintaining
its own pointer.

textEdit.dot is basically a collaboration diagram showing which objects
creates which, the parents, and which objects get swapped out if the
user calls setTextDocument().

Task-number: QTBUG-35688
Change-Id: Iac5000aa8027a75a000d4501190f3ee5f04e426e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>