qmllsutils: rename local static helper methods for completion
The naming got a bit weird in qqmllsutils for the completion methods, so
rename those methods to be easier to understand.

Previously, completion was done on "simple" items like a script
identifier expression like the name of a variable. Because the client
takes care of filtering, it does not matter to qmlls if the user
already started typing something for the identifier or not.

For more "complicated" items, like a Pragma, or the not-yet-implemented
for statement for example, completion is different depending on what
was already typed by the user:
1. If the user typed nothing, the completion can propose a snippet that can
  add the entire pragma or for-loop, for example. In this case, name
  methods that do this as suggestXXXCompletion().
2. If the user already typed something, then this something needs to be
  analyzed carefully to know what completion is currently needed. In
  this case, name methods that do this as insideXXXCompletion().

Rename the static helper methods to fit the naming scheme above, and
remove bindingsCompletions and importCompletions from the header as
both are not used anywhere except in qqmllsutils.cpp.

Change-Id: I01becc373aaba83df50d50cece7b1cc75dbf3b1e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>