disable completions for non-implemented types
Currently, implementing more completions will break unrelated tests
because of the way that QQmlLSUtils::completions() dispatches the
completion calls.

Suppress the completion for the unimplemented DomTypes: like this, the
unimplemented DomTypes do not end up with the completion of one of their
parents, but instead just do not return any completion.

Task-number: QTBUG-117445
Change-Id: I35f54d0fc5bc3482a7c6ef6a591be613e3b64aeb
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>