Rewrite QuickTests's qWaitForSignal
It doesn't use any of the features of QSignalSpy, so
it hasn't done it any favors. On top of that QSignalSpy
connects with DirectConnection which means that frequently
emitted signals on another thread may end up trying to
append another signal emission to its internal list during
destruction of the signal spy, leading to use-after-free.

The new implementation uses a small class with a slot that just sets
a boolean. We use AutoConnection so it is thread safe, and pending
emissions go away when the object is destroyed.

Fixes: QTBUG-118744
Fixes: QTBUG-118163
Pick-to: 6.6 6.5
Change-Id: Ib2ccce2369a681bfe9a2e04d2c091f9690edee49
Reviewed-by: Dimitrios Apostolou <jimis@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>