fix mismatch in QQmlLSUtils::completions
Call the helper methods using current, and not currentItem. This does
not make a difference for most of them, but that is for sure confusing
if insideIfStatement is called with a DomItem that does not represent an
if statement. Also add a comment on why this is needed.

Task-number: QTBUG-117445
Change-Id: I7af0a571100363071a47463c4ebff508f38b4237
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>