Make the resource_name more unique in qt_add_shaders call
It has been pointed out that platforms using static builds
may end up in name collisions when the resource_name for
qt_add_shaders (or qt_add_resources) clashes between a
Qt library and the application. While qtdeclarative is
not the worst offender among the Qt modules, we should
still make some of the qt_add_shaders calls more uniform
and better prefixed when it comes to the resource_name.

Pick-to: 6.6
Task-number: QTBUG-118624
Change-Id: I8fd7616f4329f652af0afa8253944d7018d9f8d1
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>