QQuickOverlay: don't reset rotation during resize if we don't have to
Overlays on windows that have an explicit content orientation set need
to reset their rotation value when the window's geometry changes, so
that items on the overlay have the correct orientation as well. This
was implemented in f1e139b9ac02313f541ed6749f05e265f4c8bf13.

However, for windows that follow the default screen orientation (i.e.
use Qt::PrimaryOrientation), we don't have to reset the rotation to
0 when the window resizes.

Fixes: QTBUG-118460
Pick-to: 6.6
Change-Id: I1a65f15000b1d8a6425773adeba127915b944581
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>