tst_qmlls_modules::buildDir: stop testing unrelated completions
The test actually tests if qmlls can find types defined in the build
folder (e.g. in qmldir etc), and tests it by checking if the build
folder types can be found in the completions.

As a drive-by, it seems to test if there are completions for property
and function keywords - but that should actually be tested in
tst_qmlls_utils, not tst_qmlls_modules. Remove the expected keywords, as
those will change anyway in the commits in the relation chain:
they will be replaced with code snippets instead, and complemented with
better tests as tst_qmlls_modules::buildDir.

Task-number: QTBUG-117445
Change-Id: I4a7f58ab9d3dd17f1374ee580e3eb1447310bbe4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>