Popup: fix fading in and out of the dimmer
When showing and hiding the overlay we go through QQmlProperty to enable
behaviors, e.g. for fading the dimmer in and out, but since we already
set the opacity to 1.0 when creating the dimmer any fade-in behavior is
ignored (as the property already has the target value).

Fix this by creating the dimmer with opacity 0, and calling showOverlay
to make it visible if the popup is already visible. In addition, store
the target opacity of the dimmer component, and set opacity to that
value instead of 1.0 to respect an explicitly set opacity on the modal
or modeless item.

Add a test to verify that we respect the opacity that is set on the
attached modal or modeless dimmer item, and that we correctly trigger
behaviors on the dimmer item. For that to work, the exit transition of
the popup itself needs to be at least as long as the behavior's
duration.

Fixes: QTBUG-118461
Pick-to: 6.6
Change-Id: I5259ea54f7e4a98bf671e8b52c26676d591a0176
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>