CMake: Fix the all_qmllint* targets for Visual Studio generators
When using the VS generator, the foo_qmllint targets were run on every
build, which is neither expected nor wanted. The reason for this is a
pecularity in MSBuild and as a reaction to this, special behavior of
CMake: add_dependencies(A B) will enable B in the default build of the
solution - even if B is not triggered by the ALL target. See upstream
CMake issue #16668 for details.

We now work around this problem by building foo_qmllint externally as
PRE_BUILD step of all_qmllint.

Pick-to: 6.5
Fixes: QTBUG-115166
Change-Id: Ie9c187edf112fc37e8eb7d1c1040fc64479b3034
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit f81945e1d79f69c27340cc2ba1a5919957eae514)