Material: respect horizontalAlignment in placeholder text
This allows placeholder text to follow the Alignment set to
the TextField or TextArea components when using Material style.
The placeholder text will float to the left, right, or center
depending on the alignment set, and the arc will be drawn properly
in case of Material.Outline container style.

Fixes: QTBUG-118856
Pick-to: 6.5
Change-Id: Ic9cede806dc2f6109e7e2c4b2b2fc960d9c6a1b6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 7e678acbb6167218d349c49436499121b2025af5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>