Fix thin text decoration with fractional scaling
With fractional scales, 1 pixel thin underlines and other
text decoration get a lot of aliasing effects. Sometimes
it will disappear completely and other times it will
fill two pixels. The rendering depends on the fractional
position of the line, so any change that moves the text
may cause the line to render at a different width, causing
instability when resizing layouts for instance.

We were using rectangle nodes for this for convenience, but
even with antialiasing enabled, that does not look correct
for single-pixel lines. Instead, we use the new curve renderer
which can also draw nice antialiased straight lines.

Fixes: QTBUG-86977
Change-Id: I1367b1e2d25d167fac7d6284cb40a652ef4c3e9a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>