TreeView: document how to implement expanding indicators
Being able to rotate the expanded/collapsed indicator
in a TreeView is assumed to be a normal request. But
it's not very intuitive to implement, since TreeView will
do a relayout when you expand or collapse a node, which
will cause existing delegates to be reused. If a delegate
has an ongoing animations at this point, it will look weird.
In order to solve that, you need to use the
TableView.onPooled/onReused signals, which is not intuitive.

Until (and if) we have have a more explicit API to support
this, at least show in the documented example how it can be
done.

Pick-to: 6.5
Change-Id: Ibf08fd8518d74ba12a4a6f7db71f3f56a8970fac
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 59daf95ac972cc93a48455835f98e79d9bff5810)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>