QML: Allow JavaScript extension types
This is what we routinely do in the builtins. In particular:
- Add the jsroot metatypes to qmltyperegistrar so that it can validate
  the JavaScript types
- Add a special class info "ExtensionIsJavaScript"
- Pass this through metatypes and qmltyperegistrar.
- Treat it like an extension type in QQmlJSScope but give it its own
  name.
- Ignore it in qmltc. We cannot express JavaScript extensions in C++.

Now qmltyperegistrar has to deal with name conflicts between JavaScript
and C++ types. It does so by always preferring the C++ type if
available. The new class info is private and we don't want to
surprisingly inject JavaScript types into people's QML modules. For our
own types we shall make sure to avoid name conflicts. There is one
prominent name conflict, 'Qt', which we really want to be resolved
towards the C++ side.

Change-Id: Iac224f4b0df2d6398fba0124f0b52d8ddfccace1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>