qmlls: change findUsages semantics
So far findUsages was based on "the same object" semantics. This means
that when findUsages is triggered for an element, we were first trying
to find the scope that this element belong to, and then return the
usages for that scope. However, there are quite a few edge cases that
this approach wouldn't work, for example the same inline component that
is instantiated in a different context. In those cases, there is no
clear definition of "the same scope".

Hence we now look for the base type where the name-to-findUsages-of is
defined first.

A few existing tests (groupPropertiesUsage, Connections) was adapted to
change. Also, try to add new tests in a separate qml files instead of a
single but large JSUsages.qml. Having a single test file is very
difficult to maintain.

Fixes: QTBUG-118712
Change-Id: I921e334723def9784ee4ca2156aa200d1db490ce
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>