QQuickPopup: fix setVisible() when a window is not yet assigned
Do not start EnterTransition for a popup in QQuickPopup::setVisible()
when the popup is not yet associated with any window. Otherwise, the
annoying warning "cannot find any window to open popup in" will be
generated and the popup won't be displayed at all, even after getting
a window.

At the same time, there is a code in QQuickPopupPrivate::setWindow()
which will start the transition for us in appropriate time.

Amends 618f34889be51f531b6226e6362f40604967b486 and
db286068f352291416d1fce9fe9953f376b74798

Fixes: QTBUG-75483
Pick-to: 6.6 6.5
Change-Id: I827f6df1a1bd7685cb293cd6b599d0a547be777e
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>