qqmllsutils: add CMake build command
Add a method QQmlLSUtils::cmakeBuildCommand() that generates the CMake
build command to run the all_qmltyperegistrations target.

This target generates the .qmltypes, .qrc, .qmldir and copies the qml
files into the build folders and generates everything that qmlls needs
from the build folder, without compiling the users code, so it should
be much faster than a complete build.

This method is hard to test (I do not believe you can call cmake just
like this in the CI), so instead just test that
cmakeBuildCommand() returns the correct cmake command.

Task-number: QTBUG-118705
Change-Id: Ia29748d8fa9b38c3d4766d250e0ee9a15da7b09e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>