qmlls: rebuild when modifying QML elements in C++ for CMake projects
newDocForOpenFile() first checks if a rebuild is needed and triggers
a rebuild before creating the Dom to incorporate the latest updates
in the C++ files defining the QML elements. This also means that
rebuilds are only triggered when the user types some code in a qml
file (to avoid constantly running builds when a user is typing C++
code).

Disable the special CMake behavior if CMake does not exist on the
current system via the m_cmakeStatus flag in QQmlCodeModel. The
assumption is that cmake exists if `cmake --version` returns with
exitcode 0.

To check if a rebuild is needed, we use QFileSystemWatch to watch
the c++ files that defines the QML elements used in the .qml files.
The QFileSystemWatch::fileChanged signal is connected to
QQmlCodeModel::onCppFileChanged to trigger builds in the next
newDocForOpenFile() call.

Finally, addFileWatches() takes a DomItem and add watches for all C++
files needed by the passed QmlFile. For testing purposes, this method is
split in two:tst_qmlls_qqmlcodemodel has tests for both halves.

The first half, fileNamesToWatch(), uses the QQmlJSTypeResolver to find
all C++ defined scopes and collects all their file names. The file name
provided by the QQmlJSScope comes from the .qmltypes, that contains a
relative path and therefore needs to be made absolute again.

For this, the second half of the method is done by
findFilePathsFromFileNames(), that uses the rootUri (populated in a
commit below this one in the relation chain) as the root source
directory to search all files having one of the names found previously
in fileNamesToWatch(). This is an overapproximation and might find
unrelated files with the same name, but that should be fine: in the
worst case, a rebuild is done for a QML module not used by the current
file.

Add tests for fileNamesToWatch, buildFolderWasCleaned,
findFilePathsFromFileNames.
This implies creating a small directory structure in the
tst_qmlls_qqmlcodemodel data test with stub C++, qml and unrelated
files, to see if the correct files are found.

Fixes: QTBUG-118708
Fixes: QTBUG-118709
Task-number: QTBUG-118705
Change-Id: Idae3c3917f4154dc8c21551f5bfc5ce050365d91
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>