QQmlDomAstCreator: do not call resourceFilesFromBuildFolders twice
Its seems that d0fcb75aab1b5a91260fb378d761257e1f2e4787 introduced a bug
where the resourceFilesFromBuildFolders was called twice during Dom
creation: the resources files were not found in the second call and the
Dom was not able to find the qmldir, qmltypes, etc.

The problem was that the current test only tests if there are linting
errors, but currently, the Dom created by the linting module is not the
Dom used by all other modules. Therefore, extend the tests to test this
second Dom by asking for completions of the type of the build/source
folder. Also change the test to be a data test.

Making the linting module of qmlls reuse the Dom that the other modules
also uses is tracked in QTBUG-119126.

Task-number: QTBUG-119126
Change-Id: I3a8db0f854512323b61b4763e8a2d84d407f46d8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>