qmlls: set the root uris of qqmlcodemodel
Currently, the root uris are not set by qmlls. These root uris point to
the root directory of the user's project. This information is made
available by the client, but was until now not used anywhere.
Populate the rootUris member of qqmlcodemodel with this information.

This is required for the 'call cmake' functionality of qmlls.

Task-number: QTBUG-118705
Change-Id: I0280ac012054efc29946e22b37b36bd6e2f69723
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>