SignalSpy: fix potential issues on disconnecting a signal
In 2e7d4ecdc59942b484159ca827f5d5dbc8787a1b and
b19ebe1d23e1f2fd334cef4ec2731ab5cc69dbd7 we added support for cases
where a signal and a property name of an object are the same, like
"pressed" for MouseArea or "opened" for Popup.

But this was missed in fd7459a60c3b396984c948f244ec0996fba56837,
which caused warnings like "SignalSpy.qml:55: TypeError: Property
'disconnect' of object true is not a function" in various tests.

So, fix this and avoid the code duplication as much as possible.

Task-number: QTBUG-118163
Pick-to: 6.6 6.5 6.2
Change-Id: Ic6dc3ffeaf470e40cd14a0800f32212a0c1defa9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>