StyleGenerator: let qrc file use relative paths
The paths in the qrc file should be relative to the location
of the qrc file itself. Otherwise you cannot bundle the
source code of the generated style into an application, since
the paths in the qrc file wil be local to a the machine
where the style generator was run.

This patch will therefore change how we generate the
qrc file so that the paths inside it are always local
to the qrc file. And we also need to ensure that
we generate the qrc file as the last thing, otherwise
files that are generated afterwards will not be included
in the qrc file.

Change-Id: I6e89f5e17b09ab760d2c5b0019d31b63cab97de0
Reviewed-by: Doris Verria <doris.verria@qt.io>