CMake: Deploy runtime dependencies of QML plugins
On Windows, we only deployed the Fooplugin.dll of a QML module Foo. The
backing library was missing. The reason was the file path heuristics we
employ that looks for '/Fooplugin.*\.dll'. That doesn't take into
account the backing library.

Now, we use the QtDeployTargets.cmake information that is generated on
project configuration and deploy the DLLs that are used at runtime.

Pick-to: 6.5 6.6
Fixes: QTBUG-117948
Change-Id: I6d26405814ca2bc66bfa308cf4d23e42cbbfeb7f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>