TestCase: make failOnWarning() functional on temp objects destruction
We need to be able to handle warnings that may occur when temporary
objects, that were created in a test function, are destroyed.

Now, with 18aa36cf878a52b8fe991392098e9445c3d3bfe3, we need only
to wait before calling finishTestDataCleanup() to ensure that the
objects marked for deletion are completely deleted.

Without this patch, we may not fail on warnings like this:

QWARN  : tst_controls::Basic::ComboBox::test_mouseHighlight() qrc:/qt-project.org/imports/QtTest/SignalSpy.qml:55: TypeError: Property 'disconnect' of object true is not a function

Pick-to: 6.6 6.5
Change-Id: I990d65005ecf88633f25dcd317846804ae9fdfd1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>