Add imageSourceSize property to QQuickDrag
Currently we can only assign an image URL to QQuickDrag to represent
the data during the drag and drop operation. We made it easier to
load arbitrary images in ac78bf7074c4aa2414b4da38db5b574bec9e4b71 :
it does not have to come from Item.grabToImage(). But the image size
depended only on the image provider's default behavior. Now we give
developers the ability to set a preferred image size when needed.

The revision of this new property and its signal is omitted because
of QTBUG-33179.

[ChangeLog][Qt Quick][Drag] The attaching type Drag has a new
imageSourceSize property, which can be used like Image.sourceSize to
scale the drag image during loading.

Fixes: QTBUG-122705
Task-number: QTBUG-120489
Task-number: QTBUG-33179
Change-Id: I92996b8f70b7ca1c9ab783d7ed9f21649d512ab9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>